Move auth store and its test to frontend/stores/
