+++ src/frontend/pages/Login.tsx
@@ -2,7 +2,7 @@ import { useState, type FormEvent } from "react";
 import { useNavigate, Navigate } from "react-router-dom";
 import { toast } from "sonner";
 import { useLogin } from "../hooks/auth";
-import { useAuthStore } from "../lib/auth";
+import { useAuthStore } from "../stores/auth";
 import { Card, CardContent, CardHeader, CardTitle } from "../components/ui/card";
 import { Input } from "../components/ui/input";
 import { Label } from "../components/ui/label";
